Add cache methods to improve the performance of the program
This commit introduces the ability to enable caching behaviours per method in a class. If the class has an IMemoryCache then it uses its own cache. If not, it generates a Partial class with a simple cache implementation. It also makes sure that the "Microsoft.Extensions.Caching.Memory" is installed and if not, a diagnostic error is thrown. Testing was performed to ensure this version meets your needs. The option of which caching method to use can make data access more efficient. This is particularly useful in situations where the same data is accessed repeatedly. (Generated by AI. :) )