Update cache handling and version in CacheSourceGenerator
This commit updates the handling of cache entries in CacheSourceGenerator by allowing MemoryCache to be nullable. It also increases the version number from 0.1.0 to 0.2.0, reflecting the changes made. The CacheSourceGenerator project file is updated by adding a package readme file, an icon URL, and modifying package release notes. A ReadMe.md was added in the 'docs' directory and unused spaces from various files have been removed.