Add custom key generator support for cache methods
This update adds the ability to use custom key generators for cache methods. The new "KeyGenerator" attribute can be used to specify a method that generates a caching key. The generator must have the same parameters as the decorated method but can return any type. The key generator can also be asynchronous, and appropriate handling is added for this case.